Improved Health View

+++ Assets/Scripts/Core/ECS/Components/Events/Player/PlayerEvents.cs
@@ -2,8 +2,8 @@
 
 namespace Core.ECS.Events.Player
 {
-    internal struct PlayerRecievedDamageEvent { public byte Value;  }
-    internal struct PlayerHealedEvent { public byte Value; }
+    internal struct PlayerRecievedDamageEvent { public int Value;  }
+    internal struct PlayerHealedEvent { public int Value; }
     internal struct EnergyReducedEvent { public float Value; }
     internal struct PlayerDiedEvent : IEcsIgnoreInFilter { }
 }
